refactor: address the code smell
