+++ frontend/components/employee/newEmployeeDialog.js
@@ -1,11 +1,8 @@
 import DialogTitle from "@mui/material/DialogTitle";
 import DialogContent from "@mui/material/DialogContent";
-import DialogActions from "@mui/material/DialogActions";
-import {Grid} from "@mui/material";
-import Button from "@mui/material/Button";
 import Dialog from "@mui/material/Dialog";
 import * as React from "react";
-import EmployeeForm from "./employee";
+import EmployeeForm from "./employeeform";
 
 const NewEmployeeDialog = (props) => {
     const {open, addEmployee, serviceList , setAddOpen} = props;
