+++ frontend/components/service/index.js
@@ -13,7 +13,7 @@ import _orderBy from "lodash/orderBy";
 
 
 const ServiceComponent = (props) => {
-    const {serviceListData, toggleBlocked, deleteService, refresh, setRefresh , employeeList} = props;
+    const {serviceListData, toggleBlocked, deleteService, refresh, setRefresh} = props;
     const [serviceListDisplay, setServiceListDisplay] = useState(serviceListData);
     const [sortServiceList, setSortServiceList] = useState(serviceListData);
     const [select, setSelect] = useState('None');
